Refactor Serilog TelegramSink to utilize BatchCycleManager with configurable rules
The Serilog TelegramSink was refactored to utilize a new BatchCycleManager. This manager is configured with a set of rules (such as BatchSizeRule and OncePerTimeRule) which specify when batches of log messages should be emitted. Also, the LogsQueueAccessContext was updated to use a ChannelReader for non-blocking, thread-safe queue operations. This change aims to improve the performance of the TelegramSink by allowing more fine-grained control over batch processing and improving thread safety.